refactor: extract simulation hook and enhance medic AI
