+++ src/components/Simulation.tsx
@@ -1,7 +1,4 @@
-import type { World as RapierWorld } from "@dimforge/rapier3d-compat";
-import { useFrame } from "@react-three/fiber";
-import { useRapier } from "@react-three/rapier";
-import { MutableRefObject, useEffect, useMemo, useRef, useState } from "react";
+import { useMemo } from "react";
 
 import { BattleWorld } from "../ecs/world";
 import { TEAM_CONFIGS } from "../lib/teamConfig";
@@ -10,10 +7,10 @@ import {
   createBattleRunner,
 } from "../runtime/simulation/battleRunner";
 import { TelemetryPort } from "../runtime/simulation/ports";
+import { useSimulation } from "../runtime/simulation/useSimulation";
 import { MatchStateMachine } from "../runtime/state/matchStateMachine";
 import { useQualitySettings } from "../state/quality/QualityManager";
 import { ObstacleVisual } from "../visuals/ObstacleVisual";
-import { recordRendererFrame } from "../visuals/rendererStats";
 import { RobotPlaceholder } from "./RobotPlaceholder";
 import { Scene } from "./Scene";
 import { SpaceStation } from "./SpaceStation";
@@ -41,8 +38,6 @@ interface SimulationProps {
   showPerfOverlay?: boolean;
 }
 
-const FRAME_SAMPLE_INTERVAL = 1 / 30;
-
 /**
  * The core simulation component.
  * Integrates the ECS world, physics, and rendering.
@@ -51,102 +46,44 @@ const FRAME_SAMPLE_INTERVAL = 1 / 30;
  * @param props - Component props.
  * @returns The Simulation component tree.
  */
-export function Simulation({
-  battleWorld,
-  matchMachine,
-  telemetry,
-  onRunnerReady,
-  obstacleFixture,
-  showPerfOverlay,
-}: SimulationProps) {
-  const runnerRef = useRef<BattleRunner | null>(null);
-
-  useEffect(() => {
-    const runner = createBattleRunner(battleWorld, {
-      seed: battleWorld.state.seed,
-      matchMachine,
-      telemetry,
-      obstacleFixture,
-    });
-
-    runnerRef.current = runner;
-    onRunnerReady?.(runner);
-
-    return () => {
-      runnerRef.current = null;
-    };
-  }, [battleWorld, matchMachine, telemetry, onRunnerReady, obstacleFixture]);
-
+export function Simulation(props: SimulationProps) {
   return (
-    <Scene showPerfOverlay={showPerfOverlay}>
-      <SimulationContent battleWorld={battleWorld} runnerRef={runnerRef} />
+    <Scene showPerfOverlay={props.showPerfOverlay}>
+      <SimulationContent {...props} />
     </Scene>
   );
-}
-
-interface SimulationContentProps {
-  battleWorld: BattleWorld;
-  runnerRef: MutableRefObject<BattleRunner | null>;
 }
 
 /**
  * Internal content of the simulation scene.
  * Handles the game loop (useFrame), rendering of entities, and physics integration.
  */
-function SimulationContent({ battleWorld, runnerRef }: SimulationContentProps) {
-  const [, setVersion] = useState(0);
-  const accumulator = useRef(0);
+function SimulationContent(props: SimulationProps) {
+  const { battleWorld } = props;
+
+  // Custom hook manages the ECS world lifecycle and system updates.
+  const { version } = useSimulation(props);
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
+  void version;
+
   const qualitySettings = useQualitySettings();
   const instancingEnabled = qualitySettings.visuals.instancing.enabled;
   const shadowsEnabled = qualitySettings.visuals.render.shadowsEnabled;
   const shadowMapSize = qualitySettings.visuals.render.shadowMapSize;
   const instanceManager = battleWorld.visuals.instanceManager;
-  const { world: rapierWorld } = useRapier();
-
-  useEffect(() => {
-    // Expose the battle world for quick debugging in DEV
-    const win = (window as unknown) as { __battleWorld?: typeof battleWorld };
-    win.__battleWorld = battleWorld;
-    return () => {
-      delete win.__battleWorld;
-    };
-  }, [battleWorld]);
-
-  // Pass Rapier world to BattleRunner for raycasting
-  // Note: Type assertion needed due to duplicate @dimforge/rapier3d-compat types
-  // between direct dependency and @react-three/rapier's bundled version
-  useEffect(() => {
-    const runner = runnerRef.current;
-    if (rapierWorld && runner) {
-      runner.setRapierWorld(rapierWorld as unknown as RapierWorld);
-    }
-    return () => {
-      // Cleanup on unmount - use captured runner reference
-      if (runner) {
-        runner.setRapierWorld(null);
-      }
-    };
-  }, [rapierWorld, runnerRef]);
-
-  useFrame((state, delta) => {
-    recordRendererFrame(state.gl, delta);
-    runnerRef.current?.step(delta);
-    accumulator.current += delta;
-    if (accumulator.current >= FRAME_SAMPLE_INTERVAL) {
-      accumulator.current = 0;
-      setVersion((value) => value + 1);
-    }
-  });
 
   const robots = battleWorld.robots.entities;
   const projectiles = battleWorld.projectiles.entities;
   const effects = battleWorld.effects.entities;
   const obstacles = battleWorld.obstacles.entities;
+
   const robotsById = useMemo(
     () => new Map(robots.map((robot) => [robot.id, robot])),
     [robots],
   );
+
   const currentTimeMs = battleWorld.state.elapsedMs;
+
   const fallbackProjectiles = useMemo(() => {
     if (!instancingEnabled) {
       return projectiles;
@@ -155,6 +92,7 @@ function SimulationContent({ battleWorld, runnerRef }: SimulationContentProps) {
       (projectile) => projectile.instanceIndex === undefined,
     );
   }, [projectiles, instancingEnabled]);
+
   const fallbackEffects = useMemo(() => {
     if (!instancingEnabled) {
       return effects;
